Add functionality for displaying the numbers clicked and displaying the answer
The users can now see live feedback on the screen for the digits that they have pressed. Once the user clicks an operator, the display clears and then updates again with the new digits they enter. After they press the equals, the display will then display the resulting number. In addition, the user is able to do new subsequent calculations after each calculation.