Add starting display and allow for user pressing equals before operator
Now the display starts with the number 0 to allow for a more user-friendly design. Also, fixed a bug where the display cleared out if the user clicked the equals button without pressing an operator button prior. If the user presses equals, the display will just remain whatever the last number they entered was.