Add ability to change operator before computation and fixed bug
Users can now choose to change their choice of operation before completing a computation. In addition, I added some script that allows for the currently chosen operator to be highlighted on the screen. Finally, I fixed a bug where the display wouldn't properly update if the second number that a user entered had a first digit that was the same as the first number.