Fix position of display div
I removed the defined position property, and instead used the align self property and set it to flex-end. I also set the display of the div to inline-block and the max-width to 330px so that the div grows with the text and the width never exceeds the width of the calculator.