Round long floating point numbers, use scientific notation for large numbers
I added code that will keep the output to no more than 11 characters. If the number is a floating point number, I round the number to a decimal place such that the amount of characters displayed is no more than 11. If the number is an integer, then I use scientific notation such that the output is no more than 10 characters. Finally, I make sure to reduce the font size appropiately for these big outputs.