Fix bugs related to zero button
In this version, I fixed a bug where the user could enter multiple leading zeros. In addition, I fixed a bug where the user couldn't start their second number off with a leading zero. Finally, I made sure that the display would keep the zero displayed if the user wanted to enter a second number that was a floating point number less than 1.