Fix bugs related to percent button
In this version, users can now enter a new number after pressing the percent button. Before, the program would chain new digits onto the outputed percent number instead of starting a new number. In addition, I added in logic for the display to be kept to a minimum amount of characters.